requirements.txt
streamlit
faiss-cpu
sentence-transformers
transformers
torch
pandas
gdown
rank_bm25
scikit-learn
numpy
accelerate>=0.26.0